wip: split signature body-file repair
